+++ modules/effects/src/main/scala/dev/profunktor/redis4cats/pipeline.scala
@@ -19,7 +19,6 @@ package dev.profunktor.redis4cats
 import cats.effect._
 import cats.effect.implicits._
 import cats.implicits._
-import dev.profunktor.redis4cats.algebra._
 import dev.profunktor.redis4cats.effect.Log
 
 object pipeline {
