+++ modules/examples/src/main/scala/dev/profunktor/redis4cats/RedisClusterTransactionsDemo.scala
@@ -18,11 +18,9 @@ package dev.profunktor.redis4cats
 
 import cats.effect.{ IO, Resource }
 import cats.implicits._
-import dev.profunktor.redis4cats.algebra.RedisCommands
+import dev.profunktor.redis4cats.Redis
 import dev.profunktor.redis4cats.connection._
-import dev.profunktor.redis4cats.connection.RedisClusterClient
 import dev.profunktor.redis4cats.effect.Log
-import dev.profunktor.redis4cats.interpreter.Redis
 import dev.profunktor.redis4cats.transactions._
 
 object RedisClusterTransactionsDemo extends LoggerIOApp {
