+++ modules/examples/src/main/scala/dev/profunktor/redis4cats/RedisGeoDemo.scala
@@ -17,11 +17,11 @@
 package dev.profunktor.redis4cats
 
 import cats.effect.{ IO, Resource }
+import dev.profunktor.redis4cats.Redis
 import dev.profunktor.redis4cats.algebra.GeoCommands
 import dev.profunktor.redis4cats.connection._
 import dev.profunktor.redis4cats.effect.Log
 import dev.profunktor.redis4cats.effects._
-import dev.profunktor.redis4cats.interpreter.Redis
 import io.lettuce.core.GeoArgs
 
 object RedisGeoDemo extends LoggerIOApp {
