+++ modules/examples/src/main/scala/dev/profunktor/redis4cats/RedisKeysDemo.scala
@@ -17,10 +17,10 @@
 package dev.profunktor.redis4cats
 
 import cats.effect.{ IO, Resource }
+import dev.profunktor.redis4cats.Redis
 import dev.profunktor.redis4cats.algebra.{ KeyCommands, StringCommands }
 import dev.profunktor.redis4cats.connection._
 import dev.profunktor.redis4cats.effect.Log
-import dev.profunktor.redis4cats.interpreter.Redis
 
 object RedisKeysDemo extends LoggerIOApp {
