+++ modules/examples/src/main/scala/dev/profunktor/redis4cats/RedisMasterReplicaStringsDemo.scala
@@ -17,11 +17,10 @@
 package dev.profunktor.redis4cats
 
 import cats.effect.{ IO, Resource }
-import dev.profunktor.redis4cats.algebra.RedisCommands
+import dev.profunktor.redis4cats.Redis
 import dev.profunktor.redis4cats.connection._
-import dev.profunktor.redis4cats.domain.ReadFrom
+import dev.profunktor.redis4cats.data.ReadFrom
 import dev.profunktor.redis4cats.effect.Log
-import dev.profunktor.redis4cats.interpreter.Redis
 
 object RedisMasterReplicaStringsDemo extends LoggerIOApp {
