+++ modules/tests/src/test/scala/dev/profunktor/redis4cats/Redis4CatsFunSuite.scala
@@ -18,11 +18,10 @@ package dev.profunktor.redis4cats
 
 import cats.effect._
 import cats.implicits._
-import dev.profunktor.redis4cats.algebra._
+import dev.profunktor.redis4cats.Redis
 import dev.profunktor.redis4cats.connection._
-import dev.profunktor.redis4cats.domain.RedisCodec
+import dev.profunktor.redis4cats.data.RedisCodec
 import dev.profunktor.redis4cats.effect.Log.NoOp._
-import dev.profunktor.redis4cats.interpreter.Redis
 import org.scalatest.BeforeAndAfterEach
 import org.scalatest.compatible.Assertion
 import org.scalatest.funsuite.AsyncFunSuite
