+++ modules/tests/src/test/scala/dev/profunktor/redis4cats/TestScenarios.scala
@@ -20,7 +20,6 @@ import java.time.Instant
 
 import cats.effect._
 import cats.implicits._
-import dev.profunktor.redis4cats.algebra._
 import dev.profunktor.redis4cats.effect.Log
 import dev.profunktor.redis4cats.effects._
 import dev.profunktor.redis4cats.hlist._
