+++ site/docs/effects/scripting.md
@@ -11,10 +11,10 @@ Purely functional interface for the [Scripting API](https://redis.io/commands#sc
 ```scala mdoc:invisible
 import cats.effect.{IO, Resource}
 import cats.implicits._
+import dev.profunktor.redis4cats.Redis
 import dev.profunktor.redis4cats.algebra.ScriptCommands
 import dev.profunktor.redis4cats.effects.ScriptOutputType
-import dev.profunktor.redis4cats.interpreter.Redis
-import dev.profunktor.redis4cats.domain._
+import dev.profunktor.redis4cats.data._
 import dev.profunktor.redis4cats.log4cats._
 import io.chrisdavenport.log4cats.Logger
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
